tests(): Add more tests for the js logic
